pyproject.toml
[project]
name = "medical-document-assistant"
version = "0.1.0"
description = "Medical Document Assistent which include a RAG system and LLM which respond throgh Streamlit fronted."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.3.23",
    "fastapi>=0.128.1",
    "langchain>=1.2.8",
    "langchain-openai>=1.1.7",
    "loguru>=0.7.3",
    "numpy>=2.4.2",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
    # MCP Server Dependencies
    "mcp>=0.9.0",
    "click>=8.0.0",
    "mcp-server-stdio<=0.1.10",
]

[project.scripts]
medical-mcp-server = "mcp_server.mcp_server:main"
